review fix styles(2)
