+++ src/components/Dashboard/Content/NewOrders/NewOrders.tsx
@@ -2,13 +2,12 @@ import React from "react";
 import {styled} from '@mui/system';
 
 const BlockNewOrders = styled('div')({
-    Width: "100%",
+    width: "15%",
     borderRadius:"6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
     margin: "17px 0px 0px 18px",
     "@media (max-width:1900px)":{
-        minWidth: "220px",
         margin: "15px 0px 0px 14px",
     }
 });
@@ -18,7 +17,6 @@ const BlockNewOrders = styled('div')({
 const NewOrders = () => {
     return (
         <BlockNewOrders>
-
         </BlockNewOrders>
     );
 };
