+++ src/components/Dashboard/Content/ProcessOrders/ProcessOrders.tsx
@@ -1,24 +1,22 @@
 import React from 'react';
 import {styled} from "@mui/system";
 
-const BlockLiveKitchen = styled('div')({
-    Width: "100%",
-    borderRadius: "6px",
+const BlockProcessOrders = styled('div')({
+    width: "15%",
+    borderRadius:"6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
     margin: "17px 0px 0px 18px",
     "@media (max-width:1900px)":{
-        minWidth: "364px",
         margin: "15px 0px 0px 14px",
     }
 });
 
-const LiveKitchen = () => {
+const ProcessOrders = () => {
     return (
-        <BlockLiveKitchen>
-            
-        </BlockLiveKitchen>
+        <BlockProcessOrders>
+        </BlockProcessOrders>
     );
 };
 
-export default LiveKitchen;
+export default ProcessOrders;
