+++ src/components/Dashboard/Content/TableRecipe/TableRecipe.tsx
@@ -1,24 +1,23 @@
 import React from 'react';
 import {styled} from "@mui/system";
 
-const BlockTableDetails = styled('div')({
-    Width: "100%",
+const BlockTableRecipe = styled('div')({
+    width: "22%",
     borderRadius: "6px",
     border: "0.5px solid #153453",
     background: "#081A2C",
     margin: "17px 0px 0px 18px",
     "@media (max-width:1900px)":{
-        minWidth: "275px",
         margin: "15px 0px 0px 14px",
     }
 });
 
-const TableDetails = () => {
+const TableRecipe = () => {
     return (
-        <BlockTableDetails>
+        <BlockTableRecipe>
 
-        </BlockTableDetails>
+        </BlockTableRecipe>
     );
 };
 
-export default TableDetails;
+export default TableRecipe;
